minor: refactor
