+++ .gitignore
@@ -1,7 +1,7 @@
 config.toml
 start-env.py
 _dev.py
-**/temp
+**/.temp
 **/env
 **/logs
 **/data
