+++ requirements.txt
@@ -1,40 +1,32 @@
-aiohttp==3.9.4
+aiohappyeyeballs==2.4.3
+aiohttp==3.11.7
 aiosignal==1.3.1
-async-timeout==4.0.2
-attrs==22.2.0
-beautifulsoup4==4.10.0
-blinker==1.6.2
-certifi==2024.7.4
-charset-normalizer==2.0.11
-click==8.1.3
-colorama==0.4.5
-colorlog==6.6.0
-commonmark==0.9.1
-discord.py==2.1.0
-Flask==2.3.2
-frozenlist==1.3.3
-html2text==2020.1.16
-idna==3.7
-IMAPClient==2.3.1
-itsdangerous==2.1.2
+attrs==24.2.0
+blinker==1.9.0
+certifi==2024.8.30
+charset-normalizer==3.4.0
+click==8.1.7
+colorama==0.4.6
+colorlog==6.9.0
+discord.py==2.4.0
+Flask==3.1.0
+frozenlist==1.5.0
+html2text==2024.2.26
+idna==3.10
+IMAPClient==3.0.1
+itsdangerous==2.2.0
 Jinja2==3.1.4
-MarkupSafe==2.1.1
-multidict==6.0.3
-numpy==1.26.4
-pandas==2.2.1
-pyfiglet==0.8.post1
-Pygments==2.15.0
-pyngrok==7.1.3
-python-dateutil==2.8.2
-pytz==2024.1
-PyYAML==6.0
-requests==2.32.0
-rich==11.1.0
-six==1.16.0
-soupsieve==2.3.1
+markdown-it-py==3.0.0
+MarkupSafe==3.0.2
+mdurl==0.1.2
+multidict==6.1.0
+propcache==0.2.0
+pyfiglet==1.0.2
+Pygments==2.18.0
+requests==2.32.3
+rich==13.9.4
 toml==0.10.2
-tzdata==2024.1
-urllib3==1.26.19
-waitress==2.1.2
-Werkzeug==3.0.6
-yarl==1.8.2
+urllib3==2.2.3
+waitress==3.0.2
+Werkzeug==3.1.3
+yarl==1.18.0
