+++ src/constants.py
@@ -1,0 +1,2 @@
+TRADINGVIEW_ALERT_EMAIL_ADDRESS = ["noreply@tradingview.com"]
+RETRY_AFTER_HEADER = "Retry-After"
